apps/telegram-bot/requirements.txt
# Telegram Bot
python-telegram-bot==21.9

# HTTP клиент для взаимодействия с Backend API
httpx==0.28.1
aiohttp==3.11.10

# Планировщик задач
apscheduler==3.10.4

# Конфигурация
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Database (опционально, если бот напрямую работает с БД)
# sqlalchemy==2.0.36
# psycopg2-binary==2.9.10

# Опциональные зависимости
# redis==5.2.0
# sentry-sdk==2.19.2

# Development dependencies (устанавливаются отдельно в dev)
# pytest==8.3.4
# pytest-asyncio==0.24.0
# pytest-cov==6.0.0
